requirements.txt
numpy>=1.20.0
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.4.0
opencv-python>=4.5.0
scikit-image>=0.18.0
peakutils>=1.3.0
Pillow>=8.0.0
# For animation saving
ffmpeg-python>=0.2.0
# GUI components
# Note: tkinter is part of standard library in Python 3, included here as a reminder
# uncomment if installing separately
# tkinter>=8.6.0